More renaming & Add ModAPI.cpp/hpp
