+++ Nexus/ModAPI.cpp
@@ -1,0 +1,306 @@
+#include "RetroEngine.hpp"
+
+#if RETRO_USE_MOD_LOADER
+std::vector<ModInfo> modList;
+int activeMod = -1;
+
+char modsPath[0x100];
+
+char modTypeNames[OBJECT_COUNT][0x40];
+char modScriptPaths[OBJECT_COUNT][0x40];
+byte modScriptFlags[OBJECT_COUNT];
+byte modObjCount = 0;
+
+char playerNames[PLAYERNAME_COUNT][0x20];
+byte playerCount = 0;
+
+#include <filesystem>
+#include <algorithm>
+
+#if RETRO_PLATFORM == RETRO_ANDROID
+namespace fs = std::__fs::filesystem;
+#else
+namespace fs = std::filesystem;
+#endif
+
+fs::path ResolvePath(fs::path given) {
+    if (given.is_relative())
+        given = fs::current_path() / given; // thanks for the weird syntax!
+
+    for (auto &p : fs::directory_iterator{ given.parent_path() }) {
+        char pbuf[0x100];
+        char gbuf[0x100];
+        auto pf   = p.path().filename();
+        auto pstr = pf.string();
+        std::transform(pstr.begin(), pstr.end(), pstr.begin(), [](char c) { return std::tolower(c); });
+        auto gf   = given.filename();
+        auto gstr = gf.string();
+        std::transform(gstr.begin(), gstr.end(), gstr.begin(), [](char c) { return std::tolower(c); });
+        if (pstr == gstr) {
+            return p.path();
+        }
+    }
+    return given; // might work might not!
+}
+
+void InitMods() {
+    modList.clear();
+
+    char modBuf[0x100];
+    sprintf(modBuf, "%smods", modsPath);
+    fs::path modPath = ResolvePath(modBuf);
+
+    if (fs::exists(modPath) && fs::is_directory(modPath)) {
+        std::string mod_config = modPath.string() + "/modconfig.ini";
+        FileIO *configFile     = fOpen(mod_config.c_str(), "r");
+        if (configFile) {
+            fClose(configFile);
+            IniParser modConfig(mod_config.c_str(), false);
+
+            for (int m = 0; m < modConfig.items.size(); ++m) {
+                bool active = false;
+                ModInfo info;
+                modConfig.GetBool("mods", modConfig.items[m].key, &active);
+                if (LoadMod(&info, modPath.string(), modConfig.items[m].key, active))
+                    modList.push_back(info);
+            }
+        }
+
+        try {
+            auto rdi = fs::directory_iterator(modPath);
+            for (auto de : rdi) {
+                if (de.is_directory()) {
+                    fs::path modDirPath = de.path();
+
+                    ModInfo info;
+
+                    std::string modDir            = modDirPath.string().c_str();
+                    const std::string mod_inifile = modDir + "/mod.ini";
+                    std::string folder            = modDirPath.filename().string();
+
+                    bool flag = true;
+                    for (int m = 0; m < modList.size(); ++m) {
+                        if (modList[m].folder == folder) {
+                            flag = false;
+                            break;
+                        }
+                    }
+
+                    if (flag) {
+                        if (LoadMod(&info, modPath.string(), modDirPath.filename().string(), false))
+                            modList.push_back(info);
+                    }
+                }
+            }
+        } catch (fs::filesystem_error fe) {
+            PrintLog("Mods Folder Scanning Error: ");
+            PrintLog(fe.what());
+        }
+    }
+}
+
+bool LoadMod(ModInfo *info, std::string modsPath, std::string folder, bool active) {
+    if (!info)
+        return false;
+
+    info->fileMap.clear();
+    info->name    = "";
+    info->desc    = "";
+    info->author  = "";
+    info->version = "";
+    info->folder  = "";
+    info->active  = false;
+
+    const std::string modDir = modsPath + "/" + folder;
+
+    FileIO *f = fOpen((modDir + "/mod.ini").c_str(), "r");
+    if (f) {
+        fClose(f);
+        IniParser modSettings((modDir + "/mod.ini").c_str(), false);
+
+        info->name    = "Unnamed Mod";
+        info->desc    = "";
+        info->author  = "Unknown Author";
+        info->version = "1.0.0";
+        info->folder  = folder;
+
+        char infoBuf[0x100];
+        // Name
+        StrCopy(infoBuf, "");
+        modSettings.GetString("", "Name", infoBuf);
+        if (!StrComp(infoBuf, ""))
+            info->name = infoBuf;
+        // Desc
+        StrCopy(infoBuf, "");
+        modSettings.GetString("", "Description", infoBuf);
+        if (!StrComp(infoBuf, ""))
+            info->desc = infoBuf;
+        // Author
+        StrCopy(infoBuf, "");
+        modSettings.GetString("", "Author", infoBuf);
+        if (!StrComp(infoBuf, ""))
+            info->author = infoBuf;
+        // Version
+        StrCopy(infoBuf, "");
+        modSettings.GetString("", "Version", infoBuf);
+        if (!StrComp(infoBuf, ""))
+            info->version = infoBuf;
+
+        info->active = active;
+
+        ScanModFolder(info);
+
+        return true;
+    }
+    return false;
+}
+
+void ScanModFolder(ModInfo *info) {
+    if (!info)
+        return;
+
+    char modBuf[0x100];
+    sprintf(modBuf, "%smods", modsPath);
+
+    fs::path modPath = ResolvePath(modBuf);
+
+    const std::string modDir = modPath.string() + "/" + info->folder;
+
+    // Check for Data/ replacements
+    fs::path dataPath = ResolvePath(modDir + "/Data");
+
+    if (fs::exists(dataPath) && fs::is_directory(dataPath)) {
+        try {
+            auto data_rdi = fs::recursive_directory_iterator(dataPath);
+            for (auto data_de : data_rdi) {
+                if (data_de.is_regular_file()) {
+                    char modBuf[0x100];
+                    StrCopy(modBuf, data_de.path().string().c_str());
+                    char folderTest[4][0x10] = {
+                        "Data/",
+                        "Data\\",
+                        "data/",
+                        "data\\",
+                    };
+                    int tokenPos = -1;
+                    for (int i = 0; i < 4; ++i) {
+                        tokenPos = FindStringToken(modBuf, folderTest[i], 1);
+                        if (tokenPos >= 0)
+                            break;
+                    }
+
+                    if (tokenPos >= 0) {
+                        char buffer[0x100];
+                        for (int i = StrLength(modBuf); i >= tokenPos; --i) {
+                            buffer[i - tokenPos] = modBuf[i] == '\\' ? '/' : modBuf[i];
+                        }
+
+                        // PrintLog(modBuf);
+                        std::string path(buffer);
+                        std::string modPath(modBuf);
+                        char pathLower[0x100];
+                        memset(pathLower, 0, sizeof(char) * 0x100);
+                        for (int c = 0; c < path.size(); ++c) {
+                            pathLower[c] = tolower(path.c_str()[c]);
+                        }
+
+                        info->fileMap.insert(std::pair<std::string, std::string>(pathLower, modBuf));
+                    }
+                }
+            }
+        } catch (fs::filesystem_error fe) {
+            PrintLog("Data Folder Scanning Error: ");
+            PrintLog(fe.what());
+        }
+    }
+
+    // Check for Scripts/ replacements
+    fs::path scriptPath = ResolvePath(modDir + "/Scripts");
+
+    if (fs::exists(scriptPath) && fs::is_directory(scriptPath)) {
+        try {
+            auto data_rdi = fs::recursive_directory_iterator(scriptPath);
+            for (auto data_de : data_rdi) {
+                if (data_de.is_regular_file()) {
+                    char modBuf[0x100];
+                    StrCopy(modBuf, data_de.path().string().c_str());
+                    char folderTest[4][0x10] = {
+                        "Scripts/",
+                        "Scripts\\",
+                        "scripts/",
+                        "scripts\\",
+                    };
+                    int tokenPos = -1;
+                    for (int i = 0; i < 4; ++i) {
+                        tokenPos = FindStringToken(modBuf, folderTest[i], 1);
+                        if (tokenPos >= 0)
+                            break;
+                    }
+
+                    if (tokenPos >= 0) {
+                        char buffer[0x100];
+                        for (int i = StrLength(modBuf); i >= tokenPos; --i) {
+                            buffer[i - tokenPos] = modBuf[i] == '\\' ? '/' : modBuf[i];
+                        }
+
+                        // PrintLog(modBuf);
+                        std::string path(buffer);
+                        std::string modPath(modBuf);
+                        char pathLower[0x100];
+                        memset(pathLower, 0, sizeof(char) * 0x100);
+                        for (int c = 0; c < path.size(); ++c) {
+                            pathLower[c] = tolower(path.c_str()[c]);
+                        }
+
+                        info->fileMap.insert(std::pair<std::string, std::string>(pathLower, modBuf));
+                    }
+                }
+            }
+        } catch (fs::filesystem_error fe) {
+            PrintLog("Script Folder Scanning Error: ");
+            PrintLog(fe.what());
+        }
+    }
+}
+
+void SaveMods() {
+    char modBuf[0x100];
+    sprintf(modBuf, "%smods", modsPath);
+    fs::path modPath = ResolvePath(modBuf);
+
+    if (fs::exists(modPath) && fs::is_directory(modPath)) {
+        std::string mod_config = modPath.string() + "/modconfig.ini";
+        IniParser modConfig;
+
+        for (int m = 0; m < modList.size(); ++m) {
+            ModInfo *info = &modList[m];
+
+            modConfig.SetBool("mods", info->folder.c_str(), info->active);
+        }
+
+        modConfig.Write(mod_config.c_str(), false);
+    }
+}
+
+void RefreshEngine() {
+    // Reload entire engine
+    Engine.LoadGameConfig("Data/Game/GameConfig.bin");
+#if RETRO_USING_SDL2
+    if (Engine.window) {
+        char gameTitle[0x40];
+        sprintf(gameTitle, "%s%s", Engine.gameWindowText, Engine.UseBinFile ? "" : " (Using Data Folder)");
+        SDL_SetWindowTitle(Engine.window, gameTitle);
+    }
+#elif RETRO_USING_SDL1
+    char gameTitle[0x40];
+    sprintf(gameTitle, "%s%s", Engine.gameWindowText, Engine.UseBinFile ? "" : " (Using Data Folder)");
+    SDL_WM_SetCaption(gameTitle, NULL);
+#endif
+
+    ReleaseGlobalSfx();
+    LoadGlobalSfx();
+
+    SaveMods();
+}
+
+#endif
