+++ Nexus/Sprite.cpp
@@ -298,7 +298,7 @@ int LoadBMPFile(const char *filePath, byte sheetID)
 
         if (gfxDataPosition >= GFXDATA_MAX) {
             gfxDataPosition = 0;
-            printLog("WARNING: Exceeded max gfx size!");
+            PrintLog("WARNING: Exceeded max gfx size!");
         }
 
         CloseFile();
@@ -368,7 +368,7 @@ int LoadGIFFile(const char *filePath, byte sheetID)
         }
         else {
             gfxDataPosition = 0;
-            printLog("WARNING: Exceeded max gfx surface size!");
+            PrintLog("WARNING: Exceeded max gfx surface size!");
         }
 
         CloseFile();
@@ -420,7 +420,7 @@ int LoadGFXFile(const char *filePath, byte sheetID)
 
         if (gfxDataPosition >= GFXDATA_MAX) {
             gfxDataPosition = 0;
-            printLog("WARNING: Exceeded max gfx size!");
+            PrintLog("WARNING: Exceeded max gfx size!");
         }
 
         CloseFile();
@@ -464,7 +464,7 @@ int LoadRSVFile(const char *filePath, byte sheetID)
 
         if (gfxDataPosition >= GFXDATA_MAX) {
             gfxDataPosition = 0;
-            printLog("WARNING: Exceeded max gfx size!");
+            PrintLog("WARNING: Exceeded max gfx size!");
         }
 
         return true;
