+++ Nexus/Userdata.cpp
@@ -6,15 +6,6 @@ char globalVariableNames[GLOBALVAR_COUNT][0x20];
 
 char gamePath[0x100];
 
-
-#if RETRO_USE_MOD_LOADER
-char modsPath[0x100];
-
-std::vector<ModInfo> modList;
-
-#include <filesystem>
-#endif
-
 void InitUserdata()
 {
     // userdata files are loaded from this directory
@@ -256,7 +247,7 @@ void InitUserdata()
 
         int nummaps = SDL_GameControllerAddMappingsFromFile(buffer);
         if (nummaps >= 0)
-            printLog("loaded %d controller mappings from '%s'\n", buffer, nummaps);
+            PrintLog("loaded %d controller mappings from '%s'\n", buffer, nummaps);
     }
 #endif
 }
@@ -342,238 +333,3 @@ void writeSettings() {
 
     ini.Write(BASE_PATH"settings.ini");
 }
-
-#if RETRO_USE_MOD_LOADER
-#include <string>
-#if RETRO_PLATFORM == RETRO_ANDROID
-namespace fs = std::__fs::filesystem;
-#else
-namespace fs = std::filesystem;
-#endif
-
-void initMods()
-{
-    modList.clear();
-
-    char modBuf[0x100];
-    sprintf(modBuf, "%smods/", modsPath);
-    fs::path modPath(modBuf);
-
-    if (fs::exists(modPath) && fs::is_directory(modPath)) {
-        std::string mod_config = modPath.string() + "/modconfig.ini";
-        FileIO *configFile     = fOpen(mod_config.c_str(), "r");
-        if (configFile) {
-            fClose(configFile);
-            IniParser modConfig(mod_config.c_str(), false);
-
-            for (int m = 0; m < modConfig.items.size(); ++m) {
-                bool active = false;
-                ModInfo info;
-                modConfig.GetBool("mods", modConfig.items[m].key, &active);
-                if (loadMod(&info, modPath.string(), modConfig.items[m].key, active))
-                    modList.push_back(info);
-            }
-        }
-
-        try {
-            auto rdi = fs::directory_iterator(modPath);
-            for (auto de : rdi) {
-                if (de.is_directory()) {
-                    fs::path modDirPath = de.path();
-
-                    ModInfo info;
-
-                    std::string modDir            = modDirPath.string().c_str();
-                    const std::string mod_inifile = modDir + "/mod.ini";
-                    std::string folder            = modDirPath.filename().string();
-
-                    bool flag = true;
-                    for (int m = 0; m < modList.size(); ++m) {
-                        if (modList[m].folder == folder) {
-                            flag = false;
-                            break;
-                        }
-                    }
-
-                    if (flag) {
-                        if (loadMod(&info, modPath.string(), modDirPath.filename().string(), false))
-                            modList.push_back(info);
-                    }
-                }
-            }
-        } catch (fs::filesystem_error fe) {
-            printLog("Mods Folder Scanning Error: ");
-            printLog(fe.what());
-        }
-    }
-}
-
-bool loadMod(ModInfo *info, std::string modsPath, std::string folder, bool active)
-{
-    if (!info)
-        return false;
-
-    info->fileMap.clear();
-    info->name    = "";
-    info->desc    = "";
-    info->author  = "";
-    info->version = "";
-    info->folder  = "";
-    info->active  = false;
-
-    const std::string modDir = modsPath + "/" + folder;
-
-    FileIO *f = fOpen((modDir + "/mod.ini").c_str(), "r");
-    if (f) {
-        fClose(f);
-        IniParser modSettings((modDir + "/mod.ini").c_str(), false);
-
-        info->name    = "Unnamed Mod";
-        info->desc    = "";
-        info->author  = "Unknown Author";
-        info->version = "1.0.0";
-        info->folder  = folder;
-
-        char infoBuf[0x100];
-        // Name
-        StrCopy(infoBuf, "");
-        modSettings.GetString("", "Name", infoBuf);
-        if (!StrComp(infoBuf, ""))
-            info->name = infoBuf;
-        // Desc
-        StrCopy(infoBuf, "");
-        modSettings.GetString("", "Description", infoBuf);
-        if (!StrComp(infoBuf, ""))
-            info->desc = infoBuf;
-        // Author
-        StrCopy(infoBuf, "");
-        modSettings.GetString("", "Author", infoBuf);
-        if (!StrComp(infoBuf, ""))
-            info->author = infoBuf;
-        // Version
-        StrCopy(infoBuf, "");
-        modSettings.GetString("", "Version", infoBuf);
-        if (!StrComp(infoBuf, ""))
-            info->version = infoBuf;
-
-        info->active = active;
-
-        // Check for Data/ replacements
-        fs::path dataPath(modDir + "/Data");
-
-        if (fs::exists(dataPath) && fs::is_directory(dataPath)) {
-            try {
-                auto data_rdi = fs::recursive_directory_iterator(dataPath);
-                for (auto data_de : data_rdi) {
-                    if (data_de.is_regular_file()) {
-                        char modBuf[0x100];
-                        StrCopy(modBuf, data_de.path().string().c_str());
-                        char folderTest[4][0x10] = {
-                            "Data/",
-                            "Data\\",
-                            "data/",
-                            "data\\",
-                        };
-                        int tokenPos = -1;
-                        for (int i = 0; i < 4; ++i) {
-                            tokenPos = FindStringToken(modBuf, folderTest[i], 1);
-                            if (tokenPos >= 0)
-                                break;
-                        }
-
-                        if (tokenPos >= 0) {
-                            char buffer[0x80];
-                            for (int i = StrLength(modBuf); i >= tokenPos; --i) {
-                                buffer[i - tokenPos] = modBuf[i] == '\\' ? '/' : modBuf[i];
-                            }
-
-                            // printLog(modBuf);
-                            std::string path(buffer);
-                            std::string modPath(modBuf);
-                            char pathLower[0x100];
-                            memset(pathLower, 0, sizeof(char) * 0x100);
-                            for (int c = 0; c < path.size(); ++c) {
-                                pathLower[c] = tolower(path.c_str()[c]);
-                            }
-
-                            info->fileMap.insert(std::pair<std::string, std::string>(pathLower, modBuf));
-                        }
-                    }
-                }
-            } catch (fs::filesystem_error fe) {
-                printLog("Data Folder Scanning Error: ");
-                printLog(fe.what());
-            }
-        }
-
-        // Check for Scripts/ replacements
-        fs::path scriptPath(modDir + "/Scripts");
-
-        if (fs::exists(scriptPath) && fs::is_directory(scriptPath)) {
-            try {
-                auto data_rdi = fs::recursive_directory_iterator(scriptPath);
-                for (auto data_de : data_rdi) {
-                    if (data_de.is_regular_file()) {
-                        char modBuf[0x100];
-                        StrCopy(modBuf, data_de.path().string().c_str());
-                        char folderTest[4][0x10] = {
-                            "Scripts/",
-                            "Scripts\\",
-                            "scripts/",
-                            "scripts\\",
-                        };
-                        int tokenPos = -1;
-                        for (int i = 0; i < 4; ++i) {
-                            tokenPos = FindStringToken(modBuf, folderTest[i], 1);
-                            if (tokenPos >= 0)
-                                break;
-                        }
-
-                        if (tokenPos >= 0) {
-                            char buffer[0x80];
-                            for (int i = StrLength(modBuf); i >= tokenPos; --i) {
-                                buffer[i - tokenPos] = modBuf[i] == '\\' ? '/' : modBuf[i];
-                            }
-
-                            // printLog(modBuf);
-                            std::string path(buffer);
-                            std::string modPath(modBuf);
-                            char pathLower[0x100];
-                            memset(pathLower, 0, sizeof(char) * 0x100);
-                            for (int c = 0; c < path.size(); ++c) {
-                                pathLower[c] = tolower(path.c_str()[c]);
-                            }
-
-                            info->fileMap.insert(std::pair<std::string, std::string>(pathLower, modBuf));
-                        }
-                    }
-                }
-            } catch (fs::filesystem_error fe) {
-                printLog("Script Folder Scanning Error: ");
-                printLog(fe.what());
-            }
-        }
-        return true;
-    }
-    return false;
-}
-void saveMods()
-{
-    char modBuf[0x100];
-    sprintf(modBuf, "%smods/", modsPath);
-    fs::path modPath(modBuf);
-
-    if (fs::exists(modPath) && fs::is_directory(modPath)) {
-        std::string mod_config = modPath.string() + "/modconfig.ini";
-        IniParser modConfig;
-
-        for (int m = 0; m < modList.size(); ++m) {
-            ModInfo *info = &modList[m];
-
-            modConfig.SetBool("mods", info->folder.c_str(), info->active);
-        }
-
-        modConfig.Write(mod_config.c_str(), false);
-    }
-}
-#endif
